Rename module ether2 -> eth2
