+++ bmv/eth2/src/test/resources/godWallet.json
