--- bmv/ether2/build.gradle
@@ -1,60 +1,0 @@
-version = '0.1.0'
-
-dependencies {
-    compileOnly("foundation.icon:javaee-api:0.9.3-SNAPSHOT")
-    implementation("org.web3j:core:4.8.1")
-    implementation("foundation.icon:javaee-scorex:$scorexVersion")
-    implementation project(':score-util')
-    implementation project(':lib')
-
-    testImplementation("org.junit.jupiter:junit-jupiter-api:5.7.2")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.7.2")
-
-    testImplementation("foundation.icon:javaee-unittest:$javaeeUnittestVersion")
-    testImplementation project(':test-lib')
-
-    testImplementation 'foundation.icon:icon-sdk:2.0.0'
-    testImplementation 'com.squareup.okhttp3:okhttp:3.11.0'
-    testImplementation 'foundation.icon:javaee-integration-test:0.9.0'
-}
-
-optimizedJar {
-    dependsOn(project(':lib').jar)
-    dependsOn(project(':score-util').jar)
-    mainClassName = 'foundation.icon.btp.bmv.ether2.BTPMessageVerifier'
-    archivesBaseName = 'bmv-ether2'
-    from {
-        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
-    } { exclude "score/*" }
-    enableDebug = debugJar
-}
-
-deployJar {
-    endpoints {
-        gangnam {
-            uri = 'https://gicon.net.solidwallet.io/api/v3'
-            nid = 7
-        }
-        local {
-            uri = scoreTest.url
-            nid = scoreTest.parseNid(scoreTest.nid)
-        }
-    }
-    keystore = scoreTest.default.keyStore
-    password = scoreTest.default.resolvedKeyPassword
-}
-
-test {
-    useJUnitPlatform {
-        if (!integrationTest) {
-            excludeTags("integration")
-        } else {
-            // use the common config files
-            systemProperty('env.props', new File('src/test/resources/env.props'))
-
-            def prefix = 'score.path.'
-            systemProperty(prefix + 'bmv-' + project.name, optimizedJar.outputJarName)
-            dependsOn optimizedJar
-        }
-    }
-}
