--- bmv/ether2/src/main/java/foundation/icon/btp/bmv/ether2/Constants.java
@@ -1,22 +1,0 @@
-/*
- * Copyright 2023 ICON Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package foundation.icon.btp.bmv.ether2;
-
-public class Constants {
-    public static final int HASH_LENGTH = 32;
-    public static final int SYNC_COMMITTEE_COUNT = 512;
-    public static final int BYTES_PER_CHUNK = 32;
-}
