+++ settings.gradle
@@ -8,7 +8,7 @@ include (
         'bmv:btpblock',
         'bmv:bridge',
         'bmv:bsc',
-        'bmv:ether2',
+        'bmv:eth2',
         'nativecoin',
         'xcall',
         'debug-bsh'
